+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 11. added test to make sure user chose at least 1 password before trying to  generate password
 12. add do while loop so user can try again if they didn't select a characer set
 13. move the get next pwd character statement into the password object - passing the character set for clarity
+14. move the random character set selector to the password object